Final_Project/backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
requests
flask-cors